Memoize theme prop in ThemeProvider
Currently, ThemeProvider does a merge (using lodash) of the theme prop
passed and the default RFTheme object. This computation is expensive but
has to happen only when the theme prop value changes. So use
React.useMemo hook to memoize the result and avoid unnecessary
re-renders (and computation of this value).