Add @emotion/is-prop-valid as optional dependency
framer/motion#298 marks @emotion/is-prop-valid as an optional
dependency. Since fictoan now depends on framer, packages that depend on
Fictoan are not aware of the optional dependency requirement, causing
yarn to throw a runtime warning on Fictoan trying to access this
package. Explicitly marking it as an optional dependency in Fictoan does
away with this warning.