Fix behaviour on paste or replace at position
Earlier, whenever a string was pasted into any of the InputFields used
under PinInputField, the values were always replaced starting from the
first PinInputField due to simply setting setValues. The ideal behaviour
though would be to paste the value only starting from the InputField
that is focused before pasting. An alternate scenario where this bug
also existed was when any of the InputFields were focused and multiple
characters were entered and the same behaviour of setting the value
starting with the first InputField occurred.

Fix this behaviour by essentially replacing the values state with the
multi-character input starting only from the focused index.